Add UFLType handler as the default handler + Move UFLType to ufl_type.py
