+++ ufl/corealg/multifunction.py
@@ -13,6 +13,7 @@ import inspect
 
 from ufl.log import error
 from ufl.core.expr import Expr
+from ufl.core.ufl_type import UFLType
 
 
 def get_num_args(function):
@@ -66,7 +67,14 @@ class MultiFunction(object):
                 for c in classobject.mro():
                     # Register classobject with handler for the first
                     # encountered superclass
-                    handler_name = c._ufl_handler_name_
+                    try:
+                        handler_name = c._ufl_handler_name_
+                    except AttributeError as attribute_error:
+                        if type(classobject) is not UFLType:
+                            raise attribute_error
+                        # Default handler name for UFL types
+                        handler_name = UFLType._ufl_handler_name_
+
                     if hasattr(self, handler_name):
                         handler_names[classobject._ufl_typecode_] = handler_name
                         break
@@ -107,5 +115,5 @@ class MultiFunction(object):
         else:
             return o._ufl_expr_reconstruct_(*ops)
 
-    # Set default behaviour for any Expr as undefined
-    expr = undefined
+    # Set default behaviour for any UFLType as undefined
+    ufl_type = undefined
